requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
camelot-py==0.9.0
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
comm==0.1.3
cryptography==36.0.2
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.1.2
frozenlist==1.3.3
ghostscript==0.7
ghp-import==2.1.0
google==3.0.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mergedeep==1.3.4
mkdocs==1.4.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.22.3
openai==0.27.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==23.0
pandas==1.4.2
parso==0.8.3
pdfminer.six==20220506
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
PyPDF2==1.27.12
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==306
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.0.2
requests==2.28.2
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
tornado==4.5.3
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.15
watchdog==2.2.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.1.2
yarl==1.8.2
